+++ src/main/java/kitchenpos/application/MenuGroupService.java
@@ -1,6 +1,6 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menugroup.MenuGroup;
 import kitchenpos.persistence.MenuGroupRepository;
 import kitchenpos.ui.dto.MenuGroupRequest;
 import org.springframework.stereotype.Service;
