+++ src/main/java/kitchenpos/domain/order/OrderLineItem.java
@@ -1,6 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.domain.menu.Menu;
 
 import javax.persistence.Entity;
 import javax.persistence.ForeignKey;
