+++ src/main/java/kitchenpos/domain/tablegroup/TableGroup.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.domain.tablegroup;
 
+import kitchenpos.domain.ordertable.OrderTable;
 import org.springframework.util.CollectionUtils;
 
 import javax.persistence.Entity;
