+++ src/main/java/kitchenpos/persistence/MenuGroupRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.persistence;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menugroup.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
