+++ src/main/java/kitchenpos/persistence/MenuRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.persistence;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.domain.menu.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
