+++ src/main/java/kitchenpos/persistence/OrderRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.persistence;
 
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.ordertable.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
