+++ src/main/java/kitchenpos/persistence/OrderTableRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.persistence;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.ordertable.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
