+++ src/main/java/kitchenpos/persistence/ProductRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.persistence;
 
-import kitchenpos.domain.Product;
+import kitchenpos.domain.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
