+++ src/main/java/kitchenpos/persistence/TableGroupRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.persistence;
 
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.tablegroup.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
