+++ src/main/java/kitchenpos/ui/dto/OrderTableResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.ui.dto;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.ordertable.OrderTable;
 
 public class OrderTableResponse {
