+++ src/test/java/kitchenpos/fixture/MenuGroupFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menugroup.MenuGroup;
 import kitchenpos.ui.dto.MenuGroupRequest;
 
 import java.util.List;
